+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [UNRELEASED]
+
+### Added
+
+### Fixed
+
+- Allow access to slices extracted from `const DerivField`.
+
+### Deprecated
+
+### Removed
+
 ## [v0.4.0] - 2025-12-10
 
 ### Added